chatbot/requirements.txt
langchain
langchain-groq
langchain-core
langchain-huggingface
langgraph
pydantic
langchain-community
python-dotenv
streamlit